feature/2.5: Added budget-plan logic for event organizers: CRUD operations + adding products/services to budget plans (with small bugs for now)
