+++ app/src/main/java/com/example/eventplannerteam22/products/data/repository/ProductRepository.kt
@@ -1,6 +1,7 @@
 package com.example.eventplannerteam22.products.data.repository
 
-import com.example.eventplannerteam22.products.data.model.ProductDTO
+import com.example.eventplannerteam22.network.ApiResult
+import com.example.eventplannerteam22.products.data.model.CreateProductDTO
 import com.example.eventplannerteam22.products.domain.Product
 import com.example.eventplannerteam22.products.domain.ProductListItem
 
@@ -8,5 +9,5 @@ interface ProductRepository {
     suspend fun getProducts(limit: Int, offset: Int): List<ProductListItem>
     suspend fun getAllProducts(limit: Int, offset: Int): List<Product>
     suspend fun getProduct(id: Int): Product
-//    suspend fun createProduct(dto: CreateProductDTO)
+    suspend fun createProduct(dto: CreateProductDTO): ApiResult<Unit>
 }
